Revert "kpatch: deprecate the replace command"
This partially reverts the following commit (except the changes in Readme,
which are OK):
-----------
commit 8e8de4718d8c13626f1010ec2851cb292a67f447
Author: Josh Poimboeuf <jpoimboe@redhat.com>
Date:   Mon Nov 16 09:38:38 2015 -0600

    kpatch: deprecate the replace command

    "kpatch replace" is complex, buggy, and probably unnecessary.  And
    upstream livepatch has nothing like it.

    Remove it from the kpatch utility, but leave the infrastructure in place
    in the patch module and the core module for now.

    Fixes: #456
-----------

'kpatch replace' is crucial when one mostly uses cumulative patches, i.e.
the patches that provide all the needed fixes in a single patch module.

It is very important to update such patches safely. That is, either the
new patch should be active after the update (if it succeeds) or the old one
should be used (if the update fails). Unloading of the old patch and then
loading the new one is more fragile because loading could fail and the
system would be left without patches at all.

The recommended workaround was to make sure the newer patch is a superset
of the old one. To be exact, the set of functions changed by the newer patch
should include all the functions changed by the older one. This is not
always possible even when one does not remove fixes from the patch.

By the way, "atomic replace" has been proposed for livepatch as well:
https://lkml.org/lkml/2018/2/6/152

The only issue with 'kpatch replace' I know about at the moment is
https://github.com/dynup/kpatch/issues/456:
"hook unload function doesn't run when module is replaced".

The current implementation of "atomic replace" for livepatch has the same
issue, by the way. It seems, it is difficult to resolve properly at the
kernel side. However, if the patches need load/unload hooks (callbacks),
we could probably design these callbacks in our cumulative patched according
to the current behaviour of Kpatch/livepatch core. Then, upgrade of our
patches would work properly. Downgrade (replacement with a previous version)
could be implemented as unload + load, it should be a rare operation anyway.