kmod/core: Do not taint the kernel
Although TAINT_LIVEPATCH and TAINT_USER are not considered errors by the
tests, add_taint() dumps the backtrace to the kernel log, which may
confuse users.

Let us not taint the kernel in kpatch.ko (livepatch is a different story).
It should be already clear from dmesg that the patch modules are loaded,
even without the taint.