Support both the older and the newer ABI of the core module
ReadyKernel currently uses kpatch 0.3.4 with a number of changes
compared to the upstream version of kpatch. Some of these changes affect
ABI of the core module (kpatch.ko), e.g. my commit 23038e4ea7e
"Allow for extension of Kpatch-related structures in the future".

The upstream kpatch also changed the ABI by changing the main structures
in commit 0bb5c106ef18 "kmod: restructure kpatch sysfs tree".

The patch modules and the core module for a given kernel must be built
using the same ABI, otherwise the patch module will fail to load, at
best.

It will be needed to rebase on top of the upstream kpatch eventually.

I'd also like not to make maintenance of kpatch tools more difficult
than it should be. Keeping several variant of the tools for different
kernel versions seems to be a real path to a maintenance nightmare.

The best solution I could come up with is to make the build process of
the core module and the patches ABI-aware and keep everything necessary
to support all ABIs we need. The variants of kmod/core/core.c,
kmod/core/kpatch.h, kmod/patch/kpatch.h kmod/patch/kpatch-patch-hook.c
are kept alongside the original sources. "-0.3" is for ABI "0.3", that
is, for the cases when it is needed to be compatible with kpatch 0.3.x
from VZ7.

This scheme should also make it easier to add patches from upstream (for
kpatch with the newer ABI) when needed, because the sources have the
same names.

The makefiles for kpatch.ko and the patch modules will now replace
core.c, kpatch.h and such with the needed versions right before the build
and will restore them to the previous state after that. A downside is
that, if one needs to debug these kernel modules, one should get the
correct sources first. Not a big deal, it seems.

So,

* kpatch.ko will now be built using the newest ABI by default. To build
it for the older ABI "0.3", pass KPATCH_ABI_COMPATIBLE=0.3 to 'make'
explicitly. The spec files for kpatch-kmod packages have been updated to
do just that.

* When building the patch modules, kpatch-build will use
the KPATCH_ABI_COMPATIBLE environment variable as the desired ABI
version. If unset, the newest ABI will be used. For example, to build
the patch modules compatible with ABI "0.3", one can do the following:
  export KPATCH_ABI_COMPATIBLE=0.3
  kpatch-build <...as usual...>
-----------

Done in the scope of https://jira.sw.ru/browse/PSBM-79803.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>