Added logging to /var/log/kpatch (v2)
The following events are logged for the patches: load, replace, unload.
Each line is for a single event. Format:
{timestamp} {event} {patch_module_name} {kernel_version} {result} [message]
  timestamp: see 'date --iso-8601=seconds'
  event: LOAD|UNLOAD|REPLACE
  kernel_version: see 'uname -r'
  result: SUCCESS|FAIL
  message: optional comment message, may provide the details of what
           happened.

The fields except 'message' do not contain spaces and are
space-separated. 'message' may contain spaces.

Example records:
2016-05-10T15:55:07+0300 LOAD kpatch_cumulative_0_4_r13 3.10.0-327.10.1.vz7.12.7 SUCCESS
2016-05-10T16:19:11+0300 UNLOAD kpatch_cumulative_0_4_r12 3.10.0-327.10.1.vz7.12.7 FAIL Failed to disable patch.

Done in the scope of https://jira.sw.ru/browse/PSBM-79803.

v2:
This is a different implementation of logging than is used in kpatch 0.3.4
for VZ7. The format of the log records is the same though.

I reused die() and globals to make the patch less invasive. It
also handles failures reported not only by load_module() and
unload_module() directly but also by the functions they call.

Saving the current context in log_event_prev helps when unload_module()
is called from load_module, for example. Luckily, unload_module() cannot
call load_module() again, so a single variable log_event_prev is enough
to keep the previous context.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>