Support "atomic replace" in the livepatch modules
Similar to "atomic replace" support in the old Kpatch core, this commit adds
parameter "replace" to the patch modules. If the module is loaded with
"replace=1", it will ask the livepatch core to replace all enabled
patches with this patch atomically. That is, either the new patch is
loaded and active after the replace operation or the older patches remain
active (if the replacement has failed).

This relies on "atomic replace" support in the kernel. At the time of
this writing, it is not merged yet. v8 of the patches is available here:
https://lkml.org/lkml/2018/2/21/518

Once these patches get into the mainline kernel, support for replace
could be added upstream to kpatch tools as well. Let us add it now for
VZ. We can revert this commit later and add the upstream implementation
once it is available.

Note that building of a livepatch module will fail right now if the
kernel does not support "atomic replace".

Done in the scope of https://jira.sw.ru/browse/PSBM-79803.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>