PC64le - do not use the patched functions as callbacks directly
It was observed by Evgenii Shatokhin in PR#755, that when the RCU
callback was called on the patched function, from unloaded livepatch
module triggered a kernel crash.

This patch implements the approach on Powerpc outlined in PR#755.
With -mcmodel=large, like any other data, function pointers are also
loaded relative to the current TOC base and are populated as relocation
entries in .toc section. Every function call/passing such function as
pointers needs to access them through .toc section + offset.

Simplest approach would be to convert the .toc + offset relocation
into a dynamic rela referring to the original function address but
the same function is also a patched function. Thus, on Powerpc we
need to append a new relocation entry into .rela.toc, duplicating the
entry used as a function pointer and converting the newly added rela
as dynamic rela.

Also, modify the function to load the new entry by modifying the
addend value to newly created dynamic rela.

Cc: Josh Poimboeuf <jpoimboe@redhat.com>
Cc: Evgenii Shatokhin <eshatokhin@virtuozzo.com>
Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>