Use atomic replace only when upgrading official cumulative patches
Atomic replacement of the binary patches results in non-obvious
behaviour w.r.t. patch/unpatch callbacks (hooks in the "old" kpatch).
That is, when patch1 is atomically replaced with patch2, the unpatch
callbacks from patch1 will not be called. It looks like this will remain
this way, see the details below:

* for the "old" kpatch: https://github.com/dynup/kpatch/issues/456
* for livepatch: https://lkml.org/lkml/2018/1/31/642

We do need atomic replace in ReadyKernel and we might need patch/unpatch
callbacks in some patches. Without the guarantee that unpatch callbacks
will be called, we'll have to be careful when using them. If these
callbacks are used to clean up something that other callbacks have done,
the last patch to unload must do all cleaning up. This should work if we
are only replacing our cumulative ReadyKernel patch with a newer
version and that version is aware of all previous versions, what is
needed to clean up there. However, things become rather difficult if we
try to downgrade the patch (the older patch is unlikely to know what
needs cleaning up for the newer patch) or when dealing with some
unifficial or non-cumulative patches (for the same reason as for
downgrade).

To sort out this mess somehow, I have rewritten "kpatch replace" command
so that atomic replace will only be used to replace an official
cumulative ReadyKernel patch with a newer version. This is the most
common case and if we decide to use callbacks in ReadyKernel patches, we
will be able to design these callbacks accordingly.

In all other cases, atomic replace is not used, the previous patches are
disabled first, then the new patch is loaded normally. The unpatch callbacks
will be called for the previous patches, then patch callbacks will be
called for the new patch. This should be OK.

So, "kpatch replace" now works as follows:

1. If one tries to use "kpatch replace" to load a non-official or a
non-cumulative patch  (this is currently detected by the names of
patch modules), all currently loaded patches are disabled first. Then
the new patch is loaded normally.

2. If one tries to load an official cumulative patch, all loaded patches
except the official cumulative patches with smaller version-release are
disabled first. Then atomic replace is used.

If the new patch fails to load, the previously disabled patches will
remain disabled. If it was not a patch upgrade operation, the system
would end up having no active binary patches at all. Far from ideal but
the risk seems acceptable. Downgrade and operations with non-official
patches should be rare anyway.

I reiterate: if the patch is upgraded, everything should work as
before. If "kpatch replace" succeeds, the new patch will be loaded and
active. Otherwise the older patch will remain loaded and active.

By the way, all this complexity indicates that we should consider using
callbacks in the most important fixes only and only if there is no other
acceptable way to provide a fix.

Done in the scope of https://jira.sw.ru/browse/PSBM-79803.