kpatch: fix 'for module in "$SYSFS"/*' loops
If no patches are loaded, "kpatch replace" failed with weird errors,
like this:

  cat: /sys/kernel/kpatch/*/enabled: No such file or directory
  kpatch: patch module kpatch_0.5.0_1.vl7.x86_64.rpm is not loaded

The root cause is that 'for module in "$SYSFS"/*' loops still do one
iteration in that case, with module == "/sys/kernel/kpatch/*",
hence '*' in some places, etc.

I fixed it the same way as it is done in the rest of the script. While
at it, I added the same fix to get_transition_patch() as well. Not sure
if it is really needed there but won't hurt.

Done in the scope of https://jira.sw.ru/browse/PSBM-79803.

Fixes: 270f5f022ab4 ("Use atomic replace only when upgrading
official cumulative patches")

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>