kpatch-build: don't die if undefined symbols are found in the patched code
Commit 7770c18f93cc ("kpatch-build: Fix the logfile checks for
'undefined'") actually broke kpatch-build, see
https://github.com/dynup/kpatch/issues/783

We hit this with the fix for CVE-2016-7097, where posix_acl_update_mode()
function was added in vmlinux and used in ext4 module.

The fix is being worked on upstream and its current version seems rather
complex.

For now, let us disable these checks so that everything would work as
before (they were effectively no-ops before).

The worst that could happen is that the patch module tried to call a
non-exported non-patches function and failed to load. We always test
loading of the patch modules before shipping them, so this is not a big
deal of a problem.

Done in the scope of https://jira.sw.ru/browse/PSBM-79803.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>