PPC64le - do not use the patched functions as callbacks directly
It was observed by Evgenii Shatokhin in PR#755, that when the RCU
callback was called on the patched function, from unloaded livepatch
module triggered a kernel crash.

This patch implements the approach on PowerPC outlined in PR#755.
With -mcmodel=large, like any other data, function pointers are also
loaded relative to the current TOC base and are populated as
relocation entries in .toc section. Every function passing a function
pointer as the argument need to load the function address through
.toc section + offset. Convert such .toc + offset relocation into
a dynamic rela, which resolves to original function address, during
module load.

Also move the comment related to nested function check, into
may_need_dynrela().

Suggested-by: Josh Poimboeuf <jpoimboe@redhat.com>
Cc: Evgenii Shatokhin <eshatokhin@virtuozzo.com>
Cc: Joe Lawrence <jdl1291@gmail.com>
Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>

https://jira.sw.ru/browse/PSBM-87678

The patch was intended for PowerPC but it did change how
create-diff-object works on x86 as well, and that should help fix
PSBM-87678.

The patch added a condition
'(rela_toc->addend == rela_toc->sym->sym.st_value)' in
function_ptr_rela(), which results in normal rela created for such
references, rather than dynrela. On x86, rela_toc->sym->sym.st_value
should be 0, because each symbol is at the offset 0 in its own section.
So function_ptr_rela() will return 0 if the a function refers to somewhere
inside this or another local function rather than to a start of a function.
Dynrelas are not meant for such relocations in the first place, normal
relas should be used. And they will be used after this patch.

Keeping the rest of the patch to simplify future rebases, etc.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>