Wait for pending RCU callbacks to complete before applying/removing the patch
If a function that schedules an RCU callback is patched as well as the
callback itself, care should be taken. It might be possible that the
original function has scheduled the callback then the patch is loaded
and the patched variant of the callback gets executed. Similar problems
may arise when unloading the patch.

This inconsistency may lead to problems.

In case we have not detected this when preparing the patch, let us wait
till all pending RCU callbacks complete before trying to apply or remove
the patch.

The small race window remains though between the end of rcu_barrier() and
the activeness safety check in register/unregister, when RCU callbacks
can still be scheduled. So, this does not fix the problem but rather
makes it less likely to show up.

N.B. Why not patch the core module instead? Because we cannot reliably
update the core module once it is loaded and there are patches depending
on it.

Done in the scope of https://jira.sw.ru/browse/PSBM-83362.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>