create-diff-object: skip dynrelas for function pointers to new functions
With #755, we started using dynrelas for function pointers.  However,
this behavior only makes sense for function pointers to existing
functions.  For function pointers to *new* functions, just use a normal
rela.

The 'function-ptr-new' unit test is from the following patch:

  https://github.com/dynup/kpatch/files/1927198/new-static-callback.patch.txt

Fixes #834.

Fixes: 495e6197508d ("kpatch-build, x86: do not use the patched functions as callbacks directly")
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>

Added in the scope of https://jira.sw.ru/browse/PSBM-87678 but it fixes
a different issue in the same code of create-diff-object.

https://github.com/dynup/kpatch/issues/834

From the description of the issue:
--------------
If a source patch adds a global function foo() that takes a callback as
an argument, and a static function bar() that is used as such callback,
create-diff-object may fail to find bar() in some cases. Namely, - if foo()
is defined in one source file, but called in another source file where bar()
is defined.

create-diff-object complains as follows:

    base.o: new function: foo
    meminfo.o: changed function: meminfo_proc_open
    meminfo.o: new function: bar
    /usr/libexec/kpatch/create-diff-object: ERROR: meminfo.o
      kpatch_create_intermediate_sections: 2485: lookup_local_symbol bar needed
      for .text.meminfo_proc_open
    ERROR: 1 error(s) encountered.
--------------

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>