kmod/core: add a fake parameter "enable"
A workaroud for https://jira.sw.ru/browse/PSBM-87522, not very elegant
but should work.

Kpatch service that loads the installed patches at boot recognizes boot
option "kpatch.enable=0". If it is set, the service does not load
patches automatically.

However, modprobe finds that boot option as well and interprets it as if
it should pass "enable=0" as a parameter to kpatch.ko kernel module when
it is being loaded. The module had no such parameter and loading failed.

The bug was there since the option was introduced but it went unnoticed
until https://jira.sw.ru/browse/PSBM-87497, because the fallback path
triggered in the older kernels (those w/o livepatch support). The core
module was eventually loaded with insmod there. No such luck with
the newer kernels.

The users might depend on "kpatch.enable=0" as a way to disable
autoloading of patches, so we cannot easily rename that parameter.
Reworking the script that loads the patches to avoid modprobe may be
doable but is extremely error-prone.

Let us use a simple solution for now. modprobe wants to pass "enabled=0"
to the kpatch core module, right? Let us add such parameter in that
module and just ignore it.

Anyway, we will eventually move to livepatch and get rid of the core
module for good.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>