'kpatch load --all' must not load more than one cumulative patch
Done in the scope of https://jira.sw.ru/browse/PSBM-95718.

It may happen (e.g. after a kernel crash during the update of a binary
patch) that 2 or more cumulative patches are installed.

'kpatch load --all' tried to load them all, which might lead to problems
if the patches touch different sets of functions.

Let us issue a warning in such cases and load only the cumulative patch
with the highest version. No restrictions are placed on the installed
non-cumulative patches.