Print the base address of the code when the patch module is loaded
Commit 3b47ac65d9e0 "cgroup, memcg: Rework css_stacks debugging." in
vzkernel allows the kernel to record the locations of the calls to
css_get/css_put.

If css_get/css_put is called from a function in a patch module and the
module is later unloaded or replaced, it can be difficult to find out
which function the recorded location belongs to.

For now, let us output the base address of the patch module's code when
the module is loaded. This way, one will be able to obtain the offsets of
the recorded locations in the "core area" of the patch module and find
the functions they belong to.

Printing the address in dmesg is not the best security practice, but it is
about a module, so leaking the address of its code should not compromise
ASLR for the kernel itself.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>