Added 'kpatch replace-with-installed' command
PSBM-104481 has shown that 16 attempts to load the new patch are sometimes
not enough when loading a new patch. If the kernel functions to be patched
are used often and for a long period of time (e.g. ext4_trim_fs() while
fstrim is running), the new patch will fail to load.

The most common case here is when the new patch is installed but the old one
(or no patch at all) is loaded.

The users may want to try loading the new patch again after the things
calm down a bit on that node. Requiring them to lookup the name of the
installed patch module via 'kpatch list' and to specify the name in
'kpatch replace' is error-prone.

I added 'kpatch replace-with-installed' command to make things easier
here. If there is a single patch installed for the currently running
kernel, this command will load it, replacing all previously loaded
patches. If no patches or 2 or more patches are installed for that
kernel, the command will report error.

Done in the scope of https://jira.sw.ru/browse/PSBM-104481.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>